feat(leaderboard): refactor polling and use ember-concurrency task
Refactor LeaderboardPoller to remove inheritance and accept course and team
instances directly, simplifying data fetching logic. Replace manual polling in
CourseLeaderboard component with an ember-concurrency task to handle leaderboard
updates more efficiently and reliably. Remove unused state and polling methods,
and integrate automatic task triggering on ActionCable messages to maintain up-
to-date leaderboard entries. These changes improve code clarity, maintainability,
and responsiveness of leaderboard data.
